requirements.txt
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.32.0
google==3.0.0
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
httplib2==0.20.4
idna==3.3
jmetalpy==1.5.5
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
plotly==5.7.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
tenacity==8.0.1
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
